+++ C++/src/3sum.hpp
@@ -5,7 +5,7 @@ using namespace std;
 
 class Solution {
 private:
-  vector<vector<int>> twoSum(int left, vector<int> &nums, int target);
+  vector<vector<int>> twoSum(size_t left, vector<int> &nums, int target);
 
 public:
   vector<vector<int>> threeSum(vector<int> &nums);
