+++ C++/src/next_permutation.cc
@@ -2,23 +2,23 @@
 #include <algorithm>
 
 void Solution::nextPermutation(vector<int> &nums) {
-  int left = nums.size() - 2;
+  size_t left = nums.size() - 2;
 
-  while (left >= 0 && nums[left] >= nums[left + 1]) {
+  while (int(left) >= 0 && nums[left] >= nums[left + 1]) {
     left--;
   }
 
   // If left is -1, it mean the current state is the last permutation
-  if (left == -1) {
+  if (int(left) == -1) {
     reverse(nums.begin(), nums.end());
   } else {
-    int right = nums.size() - 1;
+    size_t right = nums.size() - 1;
     while (nums[right] <= nums[left]) {
       right--;
     }
 
     swap(nums[left], nums[right]);
 
-    reverse(nums.begin() + left + 1, nums.end());
+    reverse(nums.begin() + long(left) + 1, nums.end());
   }
 }
